Update M3_add_topo.m
Added a check prior to interpolating topography for large grids. If the number of interpolation points is >10^7, MATLAB takes a very long time and/or crashes. Option to continue but with a warning, or automatically downsample to optimize the size to 10^7

Also, added an option to plot the UTM topo mesh rather than automatically plotting.