***Substantial changes: Updating pcolor plotting***
This update fixes some of the ugliness of pcolor model plotting. Previous versions plotted cell centers as the coordinates in pcolor because size(cell centers) is the same as size(resistivity matrix).  However, pcolor plots vertices which should be cell edges and then excludes the top row and right column of the data matrix to avoid the fencepost error. To fix this, all pcolor commands were changed to plot cell edges and the data matrix was modified to include a "dummy" row and column.

Some additional fixes were also made to make sure that axis limits were plotting correctly and station locations were plotted correctly on diagonal slices.

Finally, user_defaults now includes an option to turn grid lines on and off.