Fixed ginput bug on Mac
There is a weird bug in ginput on a Mac in which the clicked axes does not update if the crosshairs function is called. The only work-around I could find is by getting rid of the crosshairs. So now there is a special ginput function (ginput_editdata) specifically used for edit_data